Refactor out patient list section label
